Refactor - renamed to user-store
